split blueprints/robots and rails/trains
