+++ mining-and-building-tools.lua
@@ -6,6 +6,7 @@ local dirs = defines.direction
 local fa_graphics = require("graphics-and-mouse").graphics
 local fa_rails = require("rails-and-trains").rails
 local fa_belts = require("transport-belts")
+local fa_bot_logistics = require("worker-robots")
 
 local fa_building_tools = {}
